requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv

ultralytics
opencv-python
Pillow
numpy

chromadb
pandas

langchain
langchain-google-genai
google-generativeai

pydantic
tqdm
